# openapi.json
{"openapi":"3.0.1","info":{"title":"PAS API","description":"API documentation for PAS backend","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/api/issues/{issueId}/status/{status}":{"put":{"tags":["issue-controller"],"operationId":"updateIssueStatus","parameters":[{"name":"issueId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"status","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Issue"}}}}}}},"/auth/signup":{"post":{"tags":["auth-controller"],"operationId":"createUserHandler","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/auth/signin":{"post":{"tags":["auth-controller"],"operationId":"signin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/payment/{planType}":{"post":{"tags":["payment-controller"],"operationId":"createPaymentLink","parameters":[{"name":"planType","in":"path","required":true,"schema":{"type":"string","enum":["FREE","MONTHLY","ANNUALLY"]}},{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentLinkResponse"}}}}}}},"/api/messages/send":{"post":{"tags":["message-controller"],"operationId":"sendMessage","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Message"}}}}}}},"/api/issues":{"post":{"tags":["issue-controller"],"operationId":"createIssue","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Issue"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Issue"}}}}}}},"/api/issues/{issueId}/assignee/{userId}":{"post":{"tags":["issue-controller"],"operationId":"addUserToIssue","parameters":[{"name":"issueId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Issue"}}}}}}},"/api/comments":{"post":{"tags":["comment-controller"],"operationId":"createComment","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCommentRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Comment"}}}}}}},"/api/subscriptions/update":{"patch":{"tags":["subscription-controller"],"operationId":"updateUserSubscription","parameters":[{"name":"Authentication","in":"header","required":true,"schema":{"type":"string"}},{"name":"planType","in":"query","required":true,"schema":{"type":"string","enum":["FREE","MONTHLY","ANNUALLY"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Subscription"}}}}}}},"/api/users/profile":{"get":{"tags":["user-controller"],"operationId":"getUserProfile","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/subscriptions/user":{"get":{"tags":["subscription-controller"],"operationId":"getUserSubscriptions","parameters":[{"name":"Authentication","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Subscription"}}}}}}},"/api/messages/chat/{projectId}":{"get":{"tags":["message-controller"],"operationId":"getMesssageByChatId","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Message"}}}}}}}},"/api/issues/{issueId}":{"get":{"tags":["issue-controller"],"operationId":"getIssue","parameters":[{"name":"issueId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Issue"}}}}}},"delete":{"tags":["issue-controller"],"operationId":"deleteIssue","parameters":[{"name":"issueId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Issue"}}}}}}},"/api/issues/projects/{projectId}":{"get":{"tags":["issue-controller"],"operationId":"getIssuesByProjectId","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Issue"}}}}}}}},"/api/comments/{issueId}":{"get":{"tags":["comment-controller"],"operationId":"getCommentsByIssueId","parameters":[{"name":"issueId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Comment"}}}}}}}},"/api/comments/{commentId}":{"delete":{"tags":["comment-controller"],"operationId":"deleteComment","parameters":[{"name":"commentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}}}}},"components":{"schemas":{"Issue":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"assignee":{"$ref":"#/components/schemas/User"},"title":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"},"projectID":{"type":"integer","format":"int64"},"priority":{"type":"string"},"dueDate":{"type":"string","format":"date"},"tags":{"type":"array","items":{"type":"string"}}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"fullName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string","writeOnly":true},"projectSize":{"type":"integer","format":"int32"}}},"AuthResponse":{"type":"object","properties":{"jwt":{"type":"string"},"msg":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"PaymentLinkResponse":{"type":"object","properties":{"url":{"type":"string"},"id":{"type":"string"}}},"MessageRequest":{"type":"object","properties":{"senderId":{"type":"integer","format":"int64"},"projectId":{"type":"integer","format":"int64"},"content":{"type":"string"}}},"Chat":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"project":{"$ref":"#/components/schemas/Project"},"users":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}},"Message":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"content":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"chat":{"$ref":"#/components/schemas/Chat"},"sender":{"$ref":"#/components/schemas/User"}}},"Project":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"category":{"type":"string"},"tags":{"type":"array","items":{"type":"string"}},"owner":{"$ref":"#/components/schemas/User"},"issues":{"type":"array","items":{"$ref":"#/components/schemas/Issue"}},"team":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}},"CreateCommentRequest":{"type":"object","properties":{"issueId":{"type":"integer","format":"int64"},"comment":{"type":"string"}}},"Comment":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"issue":{"$ref":"#/components/schemas/Issue"},"content":{"type":"string"},"createdDateTime":{"type":"string","format":"date-time"},"user":{"$ref":"#/components/schemas/User"}}},"Subscription":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"startTime":{"type":"string","format":"date"},"endTime":{"type":"string","format":"date"},"planType":{"type":"string","enum":["FREE","MONTHLY","ANNUALLY"]},"isActive":{"type":"boolean"},"user":{"$ref":"#/components/schemas/User"}}},"MessageResponse":{"type":"object","properties":{"message":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}